refactor advisory policy evaluation detail router
